Initial work on Workspace Policy implementation for Executor
