+++ dali/pipeline/executor/executor.h
@@ -33,6 +33,7 @@
 #include "dali/pipeline/workspace/mixed_workspace.h"
 #include "dali/pipeline/workspace/support_workspace.h"
 
+#include "dali/pipeline/executor/workspace_policy.h"
 #include "dali/pipeline/graph/op_graph_verifier.h"
 #include "dali/pipeline/workspace/workspace_data_factory.h"
 
@@ -78,29 +79,6 @@ static bool HasNextStage(DALIOpType op) {
   return true;
 }
 
-struct QueueSizes {
-  QueueSizes(int output_size) : cpu_size(1), mixed_size(output_size), gpu_size(output_size) {}
-  QueueSizes(int cpu_size, int mixed_size, int gpu_size)
-      : cpu_size(cpu_size), mixed_size(mixed_size), gpu_size(gpu_size) {}
-
-  int cpu_size = 1, mixed_size = 1, gpu_size = 1;
-};
-
-struct QueueIdxs {
-  int &operator[](DALIOpType op_type) {
-    return idxs[static_cast<size_t>(op_type)];
-  }
-
-  const int &operator[](DALIOpType op_type) const {
-    return idxs[static_cast<size_t>(op_type)];
-  }
-
-  explicit QueueIdxs(int uniform_idx)
-      : idxs{uniform_idx, uniform_idx, uniform_idx, uniform_idx} {}
-
-  private:
-  std::array<int, static_cast<size_t>(DALIOpType::COUNT)> idxs = {0, 0, 0, 0};
-};
 
 struct OutputIdxs {
   int mixed;
@@ -131,7 +109,7 @@ struct OutputIdxs {
  * buffers, so that we can produce data into one while the
  * other is in use by the user.
  */
-class DLL_PUBLIC Executor {
+class DLL_PUBLIC Executor : public JIT_WS_Policy {
  public:
   using ExecutorCallback = std::function<void(void)>;
 
@@ -146,7 +124,8 @@ class DLL_PUBLIC Executor {
         event_pool_(max_num_stream),
         thread_pool_(num_thread, device_id, set_affinity),
         exec_error_(false),
-        cb_(nullptr) {
+        cb_(nullptr),
+        queue_sizes_(prefetch_queue_depth) {
     DALI_ENFORCE(batch_size_ > 0, "Batch size must be greater than 0.");
     DALI_ENFORCE(device_id >= 0, "Device id must be non-negative.");
 
@@ -181,25 +160,10 @@ class DLL_PUBLIC Executor {
   DISABLE_COPY_MOVE_ASSIGN(Executor);
 
  protected:
-  struct WorkspaceBlob {
-    workspace_store_t op_data;
+  using JIT_WS_Policy::InitializeWorkspaceStore;
+  using JIT_WS_Policy::GetWorkspace;
 
-    void Resize(int support, int cpu, int mixed, int gpu) {
-      std::get<static_cast<int>(DALIOpType::SUPPORT)>(op_data).resize(support);
-      std::get<static_cast<int>(DALIOpType::CPU)>(op_data).resize(cpu);
-      std::get<static_cast<int>(DALIOpType::MIXED)>(op_data).resize(mixed);
-      std::get<static_cast<int>(DALIOpType::GPU)>(op_data).resize(gpu);
-    }
-
-    void Clear() {
-      std::get<0>(op_data).clear();
-      std::get<1>(op_data).clear();
-      std::get<2>(op_data).clear();
-      std::get<3>(op_data).clear();
-    }
-  };
-  // TODO(klecki): do not need the blob right now
-  // vector<WorkspaceBlob> wss_;
+ 
 
 
 
@@ -221,31 +185,33 @@ class DLL_PUBLIC Executor {
 
   void SetupOutputQueuesForGraph();
 
-  template <DALIOpType op_type>
-  // workspace_t<op_type> &GetWorkspace(QueueIdxs idxs, OpPartitionId partition_idx);
-  workspace_t<op_type> GetWorkspace(QueueIdxs idxs, OpPartitionId partition_idx);
-
-  template <DALIOpType op_type>
-  // workspace_t<op_type> &GetWorkspace(QueueIdxs idxs, const OpNode &node);
-  workspace_t<op_type> GetWorkspace(QueueIdxs idxs, const OpNode &node);
-
-  template <DALIOpType op_type>
-  void SetupInputOutput(workspace_t<op_type> &ws, const OpGraph &graph, const OpNode &node,
-                        const std::vector<tensor_data_store_queue_t>& tensor_to_store_queue, const QueueIdxs idxs);
+  // template <DALIOpType op_type>
+  // // workspace_t<op_type> &GetWorkspace(QueueIdxs idxs, OpPartitionId partition_idx);
+  // workspace_t<op_type> GetWorkspace(QueueIdxs idxs, const OpGraph &graph, OpPartitionId partition_idx);
 
   // template <DALIOpType op_type>
-  // void SetupPinned(workspace_t<op_type> &ws, const OpGraph &graph, const OpNode &node,
-  //                  const QueueIdxs idxs);
+  // // workspace_t<op_type> &GetWorkspace(QueueIdxs idxs, const OpNode &node);
+  // workspace_t<op_type> GetWorkspace(QueueIdxs idxs, const OpGraph &graph, const OpNode &node);
 
-  template <DALIOpType op_type>
-  void SetupStreamsAndEvents(workspace_t<op_type> &ws, const OpGraph &graph, const OpNode &node,
-                             cudaStream_t mixed_op_stream, cudaStream_t gpu_op_stream,
-                             const std::vector<std::vector<cudaEvent_t>> &mixed_op_events,
-                             const QueueIdxs idxs);
+  // template <DALIOpType op_type>
+  // void SetupInputOutput(workspace_t<op_type> &ws, const OpGraph &graph, const OpNode &node,
+  //                       const std::vector<tensor_data_store_queue_t>& tensor_to_store_queue, const QueueIdxs idxs);
 
-  template <DALIOpType op_type>
-  workspace_t<op_type> CreateWorkspace(const OpGraph &graph, const OpNode &node,
-                                       const QueueIdxs idxs);
+  // // template <DALIOpType op_type>
+  // // void SetupPinned(workspace_t<op_type> &ws, const OpGraph &graph, const OpNode &node,
+  // //                  const QueueIdxs idxs);
+
+  // template <DALIOpType op_type>
+  // void SetupStreamsAndEvents(workspace_t<op_type> &ws, const OpGraph &graph, const OpNode &node,
+  //                            cudaStream_t mixed_op_stream, cudaStream_t gpu_op_stream,
+  //                            const std::vector<std::vector<cudaEvent_t>> &mixed_op_events,
+  //                            const QueueIdxs idxs);
+
+  // template <DALIOpType op_type>
+  // workspace_t<op_type> CreateWorkspace(const OpGraph &graph, const OpNode &node,
+  //   const std::vector<tensor_data_store_queue_t> &tensor_to_store_queue, cudaStream_t mixed_op_stream,
+  //   cudaStream_t gpu_op_stream, const std::vector<std::vector<cudaEvent_t>> &mixed_op_events,
+  //                                              const QueueIdxs idxs);
 
   class EventList {
    public:
@@ -334,6 +300,7 @@ class DLL_PUBLIC Executor {
   std::mutex errors_mutex_;
   bool exec_error_;
   ExecutorCallback cb_;
+  QueueSizes queue_sizes_;
   std::vector<tensor_data_store_queue_t> tensor_to_store_queue_;
   cudaStream_t mixed_op_stream_, gpu_op_stream_;
   // MixedOpId -> queue_idx -> cudaEvent_t
@@ -356,153 +323,27 @@ class DLL_PUBLIC Executor {
 //   using Executor::event_pool_;            \
 //   using Executor::thread_pool_
 
-template <DALIOpType op_type>
-// workspace_t<op_type> &Executor::GetWorkspace(QueueIdxs idxs, OpPartitionId partition_idx) {
-workspace_t<op_type> Executor::GetWorkspace(QueueIdxs idxs, OpPartitionId partition_idx) {
-  // auto &ws_vec = std::get<static_cast<size_t>(op_type)>(wss_[idxs[op_type]].op_data);
-  // return ws_vec[partition_idx];
-  return CreateWorkspace<op_type>(*graph_, graph_->Node(op_type, partition_idx), idxs);
-}
+// template <DALIOpType op_type>
+// // workspace_t<op_type> &Executor::GetWorkspace(QueueIdxs idxs, OpPartitionId partition_idx) {
+// workspace_t<op_type> Executor::GetWorkspace(QueueIdxs idxs, const OpGraph &graph, OpPartitionId partition_idx) {
+//   // auto &ws_vec = std::get<static_cast<size_t>(op_type)>(wss_[idxs[op_type]].op_data);
+//   // return ws_vec[partition_idx];
+//   return CreateWorkspace<op_type>(*graph_, graph_->Node(op_type, partition_idx), 
+//       tensor_to_store_queue_, mixed_op_stream_, gpu_op_stream_, mixed_op_events_, idxs);
+// }
 
-template <DALIOpType op_type>
-// workspace_t<op_type> &Executor::GetWorkspace(QueueIdxs idxs, const OpNode &node) {
-workspace_t<op_type> Executor::GetWorkspace(QueueIdxs idxs, const OpNode &node) {
-  DALI_ENFORCE(node.op_type == op_type,
-               "Wrong variant of method selected. DALIOpType does not match.");
-  // auto &ws_vec = std::get<static_cast<size_t>(op_type)>(wss_[idxs[op_type]].op_data);
-  // return ws_vec[node.partition_index];
-  return  CreateWorkspace<op_type>(*graph_, node, idxs);
-}
+// template <DALIOpType op_type>
+// // workspace_t<op_type> &Executor::GetWorkspace(QueueIdxs idxs, const OpNode &node) {
+// workspace_t<op_type> Executor::GetWorkspace(QueueIdxs idxs, const OpGraph &graph, const OpNode &node) {
+//   DALI_ENFORCE(node.op_type == op_type,
+//                "Wrong variant of method selected. DALIOpType does not match.");
+//   // auto &ws_vec = std::get<static_cast<size_t>(op_type)>(wss_[idxs[op_type]].op_data);
+//   // return ws_vec[node.partition_index];
+//   return  CreateWorkspace<op_type>(*graph_, node, tensor_to_store_queue_, mixed_op_stream_,
+//       gpu_op_stream_, mixed_op_events_, idxs);
+// }
 
-// We instantiate the operation of adding the input only for parent op_type and device
-// that are specifically allowed
-// We always use queue_idx = 0 if give queue has only one element -> it is not queued
-template <DALIOpType op_type, DALIOpType producer_type, DALITensorDevice device>
-en_if_t<allows_op_input<op_type>(producer_type) && allows_tensor_input<op_type>(device)> add_input(
-    workspace_t<op_type> &ws, const tensor_data_store_queue_t &storage, int queue_idx = 0) {
-  auto &queue = get_queue<producer_type, device>(storage);
-  DALI_ENFORCE(!queue.IsBuffered() || queue_idx < static_cast<int>(queue.size()),
-               "Backing Tensor store queue has not enough elements.");
-  auto tensor = queue[queue_idx];
-  ws.AddInput(tensor);
-}
 
-// If parent op_type or device is not allowed this is a no-op
-template <DALIOpType op_type, DALIOpType producer_type, DALITensorDevice device>
-en_if_t<!allows_op_input<op_type>(producer_type) || !allows_tensor_input<op_type>(device)>
-add_input(workspace_t<op_type>, const tensor_data_store_queue_t, int = 0) {}
-
-// This will be only used for allowed ones (TODO(klecki) with exception of the device)
-template <DALIOpType op_type, DALITensorDevice device>
-void add_output(workspace_t<op_type> &ws, const tensor_data_store_queue_t &storage,
-                int queue_idx = 0) {
-  auto &queue = get_queue<op_type, device>(storage);
-  DALI_ENFORCE(!queue.IsBuffered() || queue_idx < static_cast<int>(queue.size()),
-               "Backing Tensor store queue has not enough elements.");
-  auto tensor = queue[queue_idx];
-  ws.AddOutput(tensor);
-}
-
-// TODO(klecki): should we move this to OpNode, and make it subclasses for
-// all DALIOpTypes -> implement `add_input` and add_output for all of the subclasses
-// as well with later operations
-template <DALIOpType op_type>
-void Executor::SetupInputOutput(
-    workspace_t<op_type> &ws, const OpGraph &graph, const OpNode &node,
-    const std::vector<tensor_data_store_queue_t> &tensor_to_store_queue, const QueueIdxs idxs) {
-  for (int j = 0; j < node.spec.NumRegularInput(); ++j) {
-    auto tid = node.parent_tensors[j];
-    auto &parent_node = graph.Node(graph.Tensor(tid).producer_edge.node);
-    auto parent_op_type = parent_node.op_type;
-    auto tensor_device = graph.Tensor(tid).producer_edge.storage_device;
-
-    VALUE_SWITCH(parent_op_type, parent_op_static,
-        (DALIOpType::GPU, DALIOpType::CPU, DALIOpType::MIXED, DALIOpType::SUPPORT),
-    (
-      VALUE_SWITCH(tensor_device, device_static, (DALITensorDevice::CPU, DALITensorDevice::GPU),
-      (
-        add_input<op_type, parent_op_static, device_static>(ws,
-                                                            tensor_to_store_queue[tid],
-                                                            idxs[parent_op_static]);
-      ), DALI_FAIL("Unexpected device"))  // NOLINT(whitespace/parens)
-    ), DALI_FAIL("Unexpected op_type"));  // NOLINT(whitespace/parens)
-  }
-
-  // Argument inputs can be handled genericaly
-  for (const auto &arg_pair : node.spec.ArgumentInputs()) {
-    // Get each argument input and add them to this op's workspace.
-    auto input_index = arg_pair.second;
-    auto tid = node.parent_tensors[input_index];
-    auto &queue =
-        get_queue<DALIOpType::SUPPORT, DALITensorDevice::CPU>(tensor_to_store_queue[tid]);
-    auto tensor = queue[idxs[DALIOpType::MIXED]];  // TODO(klecki): check queueueueueuing
-    ws.AddArgumentInput(tensor, arg_pair.first);
-  }
-
-  for (int j = 0; j < node.spec.NumOutput(); ++j) {
-    auto tid = node.children_tensors[j];
-    auto tensor_device = graph.Tensor(tid).producer_edge.storage_device;
-    VALUE_SWITCH(tensor_device, device_static, (DALITensorDevice::CPU, DALITensorDevice::GPU),
-    (
-      add_output<op_type, device_static>(ws, tensor_to_store_queue[tid], idxs[op_type]);
-    ), DALI_FAIL("Unexpected device"));  // NOLINT(whitespace/parens)
-  }
-}
-
-template <DALIOpType op_type>
-void Executor::SetupStreamsAndEvents(workspace_t<op_type> &ws, const OpGraph &graph,
-                                     const OpNode &node, cudaStream_t mixed_op_stream,
-                                     cudaStream_t gpu_op_stream,
-                                     const std::vector<std::vector<cudaEvent_t>> &mixed_op_events,
-                                     const QueueIdxs idxs) {
-  /* No-op if we are not Mixed or GPU */
-}
-
-template <>
-inline void Executor::SetupStreamsAndEvents<DALIOpType::MIXED>(
-    MixedWorkspace &ws, const OpGraph &graph, const OpNode &node, cudaStream_t mixed_op_stream,
-    cudaStream_t gpu_op_stream, const std::vector<std::vector<cudaEvent_t>> &mixed_op_events,
-    const QueueIdxs idxs) {
-  // We assign unique stream to mixed ops.
-  // This ensures that we won't have false dependencies
-  // between mixed ops and the previous iterations
-  // gpu ops.
-  ws.set_stream(mixed_op_stream);
-  ws.set_event(mixed_op_events[node.partition_index][idxs[DALIOpType::MIXED]]);
-}
-
-template <>
-inline void Executor::SetupStreamsAndEvents<DALIOpType::GPU>(
-    DeviceWorkspace &ws, const OpGraph &graph, const OpNode &node, cudaStream_t mixed_op_stream,
-    cudaStream_t gpu_op_stream, const std::vector<std::vector<cudaEvent_t>> &mixed_op_events,
-    const QueueIdxs idxs) {
-  // I/O pipeline is always going to be launched alongside
-  // some other GPU work (like DL training).
-  // Therefore it is not necessary to use more than
-  // 1 stream for GPU ops, even though we may not fill
-  // the whole GPU with just I/O pipeline kernels
-  // by doing so.
-  ws.set_stream(gpu_op_stream);
-  for (const auto &p : node.parents) {
-    if (graph.NodeType(p) == DALIOpType::MIXED) {
-      const auto &parent_op = graph.Node(p);
-      // We need to block on this op's event to
-      // make sure that we respect the dependency
-      ws.AddParentEvent(mixed_op_events[parent_op.partition_index][idxs[DALIOpType::MIXED]]);
-    }
-  }
-}
-
-template <DALIOpType op_type>
-workspace_t<op_type> Executor::CreateWorkspace(const OpGraph &graph, const OpNode &node,
-                                               const QueueIdxs idxs) {
-  workspace_t<op_type> ws;
-  SetupInputOutput<op_type>(ws, graph, node, tensor_to_store_queue_, idxs);
-  // SetupPinned<op_type>(ws, graph, node, idxs);
-  SetupStreamsAndEvents<op_type>(ws, graph, node, mixed_op_stream_, gpu_op_stream_,
-                                 mixed_op_events_, idxs);
-  return ws;
-}
 
 }  // namespace dali
